refactor

+++ Projekt/project/utils/levenshteindistance.cpp
@@ -6,16 +6,16 @@ LevenshteinDistance::LevenshteinDistance()
 }
 float LevenshteinDistance::compare(std::string s1, std::string s2)
 {
-    size_t n = s1.size(), m = s2.size();
-    int d[n+1][m+1];
+    unsigned long long n = s1.size(), m = s2.size();
+    unsigned long long d[n+1][m+1];
     d[0][0]=0;
-    for(size_t i = 1; i <= n; ++i)
+    for(unsigned long long i = 1; i <= n; ++i)
         d[i][0] = i;
-    for(size_t j = 1; j <= m; ++j)
+    for(unsigned long long j = 1; j <= m; ++j)
         d[0][j] = j;
     char cost=0;
-    for(size_t i = 1; i <= n; ++i)
-        for(size_t j = 1; j <= m; ++j)
+    for(unsigned long long i = 1; i <= n; ++i)
+        for(unsigned long long j = 1; j <= m; ++j)
         {
             if( s1[i-1]==s2[j-1])
                 cost=0;
